New Function for Feitoria, Bug Fixes
A new function is called for improvements that can only be built in
city-state territory (only luxury siphoning ones are currently
supported), various bug fixes concerning accurate happiness value
calculation from a resource, checking for conflicting missions, and
others. Also, the new strategic value function for plots is now used
when a plot is outside of city radius, so fort building is possible.