EarlyExpansion Tweaks
Being alone on a continent now forces the AI to always run the Early
Expansion strategy instead of adding expansion flavor to the ideal city
limit. Also, even if the player's starting continent passes the tile
ownership checks, as long as there is a possible settling plot on the
continent that is near one of the player's existing cities and the
player has not reached their ideal city target, the strategy will still
be active.