Fast Comparisons, FPTP Tweaks, Score Victory Fix
FASTMAX() and FASTMIN() will now be used instead of MAX() and MIN(), the
scoring of happiness from a resource is now more accurate and granular,
EstimatedEndTurn() is used instead of MaxTurns() so that functions still
work if score victory is disabled, and the AI will now attempt to vote
*against* candidates if it doesn't think it will win a candidate
proposal (eg. Choose Host, Diplomatic Victory).