Minor GetRangePlusMoveToshot() Optimization
Instead of subtracting 1 from the move and shoot range of a unit if it
needed setting up, the amount of setups required is subtracted from the
move and shoot range. This not only expands the featureset to properly
support units with more than one setup required before firing, but is
also an optimization because two less conditionals are called (since
isMustSetUpToRangedAttack() fires a conditional as well).