Tactical Unit Strength Reduction With Damage Fix
CvTacticalUnit now stores the effective health influence on unit
strength instead of the raw health percent. Eg. a half-health unit's
CvTacticalUnit health percent will say it is at 75% health, because
that's how much its damage is lowered when it is at 50% health. This is
primarily because CvTacticalUnit's sort function multiplies health
percent by damage to determine which units it wants to prioritize.