Fixed Parent Node Never Caching
Fixed the fact that the shortcut borrowed from the Community Patch that
returns true much sooner doesn't actually cache values for the parent
node, which can cause problems if cached values of the parent node are
checked (since the values returned would be the default ones, e.g.
bCanEnterTerrain would be false for the parent node).