Extra OtherPlot Helpers, Fixed Bad Damage Modifiers
Added a large slew of extra arguments for combat strength/damage
calculators that improve calculation accuracy for future planning (eg.
can now get the ranged damage a unit would deal if it were in a plot
other than its current one firing at a target who is not yet in the plot
to be fired at).

Fixed a few badly set up combat damage modifiers. For example, bonus
combat strength awarded in friendly or non-friendly territory was
previously not applied to units defending from ranged attacks and the
bonus was awarded to ranged attackers based on their target plot instead
of the plot in which they were standing.

Slight modifications to attack-and-move functions to improve
consistency, slightly speed up checks, and improve code readability.