Assortment of Extras, Optimizations
Can now get the nth best interceptor of a unit (useful if we already
know that the previous n-1 interceptors will be neutralized before this
unit attacks).

Fixed IsStackedGreatGeneral() not detecting Great Admirals.

GetMovablePlotListOpt() now stores the proper pathfinder at the
beginning of the function instead of constantly checking which one to
call during the loop.