DangerPlots Update, Optimizations
If a fighter already knows it needs to move based on city/carrier
health, it no longer goes through the computationally expensive check of
whether it is devoid of nearby targets.

GetNumEnemyAirUnitsInRange() now has an overload where it takes int
arguments as references for counting fighters and bombers, so it only
needs to be called once if both numbers are required.

GetBestAirSweepTarget() and GetNumEnemyAirUnitsInRange() now use the
updated DangerPlots to return more accurate info.