Queued Attacks Removed, ExecuteAttack() Restructure
Queued attacks are now removed from the AI system: it will now launch
attacks individually, responding to the actual results of attack rolls
like a human player would (instead of relying on the expected damage
output).

ExecuteAttack() is being restructured as a result of removed queued
attacks and remade Danger Plots, though it's still WIP. For example, the
function will now have two versions, one for attacking a unit and an
overload for attacking a city.