Great Work Swapping Remade
The algorithm for probing and executing the allocation of great works
within an AI's empire has been restructured to the point of almost being
remade. Along with performance improvements, the remake has fixed the
following bugs:
- The AI did not prioritize between Great Works (i.e. it didn't try to
use its more common ones for slots before using its rarer ones)
- The AI did not swap Music Great Works
- The AI did not prioritize Great Work slots of buildings that award
Unit XP if the Great Work slot is filled
- The AI would prioritize filling Museums inefficiently over filling
some other buildings efficiently
- The AI would only swap Great Works with players if it was one short of
the necessary types of Great Works for a bonus
- The AI was not able to process equal Art-Artifact bonus cases if the
amount of slots was not a multiple of 4.
- The AI preferred to fill wonders and museums in cities in danger of
being taken next turn over filling the buildings of cities not in danger
but not able to provide any more theming bonuses.