Better Roads Distance Adjustment
Instead of dividing cutoff distance by a number (usually up to 3) to
adjust a distance filter for roads, movement distance is multiplied by
the number of moves that unit would have on the best road type the
player can build. The bulk of the calculation is done once at the
beginning of the turn, so fetching the multiplier should be fast.