Roads Support, Pathfinder Caching
The range on plots for which a unit's danger value is calculated now
factors in possible road movement.

After calculating whether or not a unit can reach a plot via ignore
units pathfinder, the result is stored away in a vector of plots-boolean
pairs to be used by other players' DangerPlots calculations
(hypothetical performance boost, still need to test).