Housecleaning, DangerPlots and Misc. Fixes
Removed some unnecessary comments.

DangerPlots values now properly calculate healing actions, including
healing after possibly defeating an enemy unit. In addition, DangerPlots
will now properly consider the one extra turn of fortification for units
that are planning on fortifying to heal.

The instant heal promotion now properly makes use of the remade
DangerPlots system.

Units that attack onto a tile that is visible to an AI player will now
be considered when calculating the next turn's DangerPlots for that AI
player, even if the unit is not otherwise visible to the AI player. This
effectively mimics human "intuition".

Fixed move-and-shoot logic not applying properly to ranged units that
need to set up before attacking.