Various DealAI and DiplomacyAI Changes
Altered the way resources, cities, open borders agreements, defensive
pacts, and research agreements are scored by the AI that handles
diplomatic deals.

Altered the way bribing and being bribed to declare war on a 3rd player
is scored, with the AI now also considering possible further wars this
may trigger and possibly even bribing players they don't like to go into
a war that they would probably lose.

If an AI wants to conquer its opponent, it will now be willing to accept
peace offers that would force its opponent to give up a large amount of
their cities.

The minimum amount of turns that the AI will wait before even
considering peace with a human player now scales with game speed.

Damage inflicted during a war now has a reverse effect on the peace
deals the AI will offer and willing to accept: if damage is
disproportionately dealt to one side during a war, the AI will expect
the other side to pay reparations in the peace deal.