Constification, DangerPlots Expansion and Fixes
Altered various parameters and functions that could be const to be
const.

Altered the way city garrison overrides are handled to cleaner solution.
As a result, the city can now also assume the health of the garrisoning
unit as well.

The DangerPlots attack target argument is now a plot pointer instead of
a unit pointer, so it can also process cities. DangerPlots' attack
target argument is fully fleshed out and will return the total damage
delta on the unit after it has attacked; this includes situations where
the unit changes plot (eg. it captures a city from an attack or advances
onto the enemy's tile) and when it heals after killing a unit.
Note that there still are some omissions (eg. if the attack target is
not killed when the unit advances, such as for Winged Hussars, the new
danger calculation will ignore the withdrawn unit).