Constification, Code Reorganization, Fixes
Altered some arguments and functions to be const.

Fixed the fact that Great Admirals were always ignored by the
TacticalAI.

Units that are blocking another unit will now attempt to move more than
one tile away so long as they would have movement remaining afterwards.

Completely revamped the way damage estimates are calculated so that the
AI only relies on them for stance estimates and no longer overestimates
its strength. Damage estimates are now overall more accurate as well.

Completely reorganized and rewrote some parts of ExecuteAttack() to be a
lot more portable: separate functions now exist for attacking with an
individual unit (ranged or melee) as well as for allocating
target-adjacent tiles to melee units. Overall, ExecuteAttack()'s code is
a lot cleaner and future changes to its code will be much easier to
implement.

Fixed oversights in ProcessAirUnitsInAttack() that resulted in the AI
not using Fighters and Bombers properly.

Great Generals (and Admirals!) will now properly be called up to improve
the combat strength of units who will participate in attacks. This
extends outside of armies as well.

A combination of remade DangerPlots' functionality expansion, a function
that returns whether a unit could survive if it does not attack this
turn (and possibly retreat), removed queued attacks, and a reorganized
ExecuteAttacks() means that the remade DangerPlots system is now fully
used when attacking targets.

Miscellaneous optimizations and performance improvements.