Plot, Settling, and Movement Tweaks and Fixes
Updated the radar fix to work a smidgeon faster than before (removed a
useless check, made the check happen once and return NULL instead of
happening for every unit in the plot).

When calculating the yields of a resource near a potential settling
plot, the AI now considers the total amount of added yield from the
resource's improvement, not just the extra yield that the improvement
gives to the resource. This represents a large enough delta in fertility
that a certain global constant, BUILD_ON_RESOURCE_PERCENT, has been set
to 0 (from -50).

Implemented new XML constants for the ring multiplier for yields on the
city plot, culture yield multiplier for determining starting locations,
and culture yield multiplier for determining possible settling spots.

Implemented yield-specific ring dividers for settling: when looking for
plots to settle, the value of certain yields and factors is divided by a
certain amount depending on which ring the tile is located in. For
example, Food's Ring 3 divider is 2, meaning that food yield in the 3rd
ring around a plot is valued at half as much as usual, while
Production's Ring 3 divider is 1, meaning production yield in the 3rd
ring around a plot is not changed. This is mean to stop the AI from
overemphasizing plots that were high on food, but only in the faraway
3rd ring.

The AI no longer applies a flat fertility boost to plots within 5 tiles
to a natural wonder: the fertility boost of such plots is boosted
naturally through the wonder's plot yields and happiness bonus.

Before industrial era, the AI will automatically assume that tiles with
fresh water, at least 1 food yield, and no resources produce 1 extra
food to factor in the tile's Civil Service benefit.

The starting location weight of features' now extends to tiles more than
just one distance away, but at a rate that diminishes with distance (1/2
for 2 tiles away, 1/3 for 3 tiles away).

When considering plots to settle, the AI will run through all buildings
with special plot requirements that it can build at a certain plot and
add their average flavor value weighted by the leader's flavor values
into the plot's fertility value. Buildings that will only become
available in future eras have their weight diminished by the difference
in eras between the AI's current era and the era of the tech(s) that
unlock the building. Wonders' values are halved, while unique buildings'
values are doubled.

Added a few new helper functions and helper arguments to existing
functions that feed into new plot settling functionality.

Migrated CvCity::IsValidBuildingLocation() to CvPlot so that it does not
need to be called by an existing city to check whether a building can be
built in a (potential) city on a given plot. It also made sense because
none of that function's checks actually referenced the city itself.

Iroquois' Unique Ability is now fixed in that it now also treats
friendly forests as roads when moving from them to a road tile that is
not a friendly forest.