Added Possible Refund of Incompatible Promotions
The following feature has been added to the code, but is disabled for
now because it represents a fairly large gameplay change:

When a unit with certain promotions is upgraded to another unit who
cannot make use of those promotions (eg. Chariot Archer with Barrage
upgraded to Knight), the new unit gets those promotions refunded and the
player can reassign promotions for that unit at the beginning of the
next turn.