Citizen Allocation on Growth, Excess Food Handling
When a city has to assign a population to a plot or a specialist as a
result of having grown from the last turn, it will ignore food yield for
the initial population placement, then reallocate the new citizen
normally once all the end-of-turn yields are calculated (production,
gold, science, culture, faith, GPP).

When scoring tiles or specialist slots at non-positive excess food, but
the tile or specialist slot would add enough food to produce positive
excess food, only the non-positive part of the future food yield has its
score multiplied by 8, the positive part has its score handled as if the
city has positive excess food.

Extended the unhardcoding of population food consumption in citizen
automation to food yield multipliers when on partial or full food focus.