Citizen Automation Tweaks and Fixes
If the city is producing a unit that uses food for production, the
citizen manager allocates food as if it were the appropriate amount of
production.

City bonuses to the output of certain yields is considered for citizen
management. For example, if a city has a Market (+25% gold), the value
of gold for citizen management is increased by 25%.

If a city has growth rate modifiers, the extra food is considered where
applicable.

Avoid growth is only forced on from happiness if the extra population
would lower happiness.

The Avoid Growth checkbox is no longer ignored if happiness is disabled.

Fixed Food Focus when Avoid Growth is enabled to no longer lower the
value of food yield when excess food per turn is negative.