DangerPlots Updates
When a unit is planning to attack a tile, the danger values calculated
for it consider it to not have a fortification bonus.

Units are now checked whether they can actually attack the tile before
their damage is added into the overall danger value.

New functions were added that will allow for the updating of DangerPlots
information on a plot-by-plot basis in the future.