Danger Pathfinder, Ghostfinder, Misc. Optimizations
Added support for TurnsToReachTarget() to use the dedicated DangerPlots
pathfinder.

Added a Ghostfinder, which is a way for the AI to check how long it
would take the average unit it creates to move from one tile to another:
since the ghost, ie. the unit being pathfinded, is created and destroyed
each call, any free promotions applied to units (eg. embarkation) are
properly applied to the ghost.