Per-city WonderProductionAI, Various Tweaks
Added support for the fact that WonderProductionAI is now city-based
instead of player-based.

Civilian units will move to safety instead of patrolling.

Road benefits are properly applied when looking for a safe plot to move
a civilian to.

AI Great Engineers can move and hurry production on the same turn.

The function that searches for a safe plot to move a civilian to now
makes use of the A* turn limiter optimization.