Ghostfinder, Various Tweaks
Added support for the Ghostfinder, a pathfinder that simulates how long
it would take for a newly created unit to go from one plot to another.

Addressed the fact that WonderProductionAI is city-based instead of
player-based.

Added a new function to check whether a resource is useful for a player
(eg. horses are no longer useful in Modern Era). This is used in the
DealAI.