Spacing, Coastal, and Happiness Value Changes
The AI will no longer count happiness contribution from resources that
are already in its borders but are not improved.

Fertility modifiers from settling on coast while already having an
affinity for coast is now both continuous and compounds less.

The previous system of modifying fertility based on spacing between the
player's cities and a "sweet spot" system is now replaced with a
modifier based on a beta distribution whose shape is determined by the
player's flavors.

The previous system of modifying fertility based on closeness to the
nearest enemy city and the boldness of the civ has been changed to a
bell-curve multiplier based on defense flavor and ghostfinding to the
nearest allied city as well as a bell-curve divider based on boldness
and ghostfinding to the nearest enemy city.