Various Tweaks and Fixes
Observers can now see the tile yields of cities.

Tweaked the order that internal objects are initialized to work with the
new Lua flavor scripting system.

Fixed an exploit where a production modifier to a certain building would
also apply to production sent into overflow, and thus could be used to
build other things faster. For example, Piety opener could be exploited
with the constant building and selling of Shrines to essentially double
a city's hammer output.

City-states will no longer attempt to grow towards luxuries that are
outside their city radius.

Added support for the fact that WonderProductionAI is now city-based
instead of player-based.