Various Tweaks and Fixes
Implemented a new initialization function specifically for "ghosts" used
in the ghostfinder.

Movement range increase from roads in the range to move and shoot
function no longer also applies an increase to the movement cost of
setting up to ranged attack.

The intercept and evasion fix has now been retooled to simply be a fix
to allow for promotions that affect air combat for units that are not
eligible for air combat if the promotion also enables air combat for the
unit (eg. air sweep combat modifier is allowed for non-fighters if the
promotion also makes the unit capable of air sweeps).

The function that searches for plots a unit can move and shoot from now
supports the ability to use the Danger pathfinder.