Unified Yield Value Evaluation, Pruning
The value of a set of yields is now centralized into one public function
in CityCitizens. This function is used to determine the value of a
specialist slot and the value of a plot to work, as well as other
possibilities that did not exist before.

The value of trade routes is now completely unbiased and uses the newly
centralized yield value determination function to determine the value of
trade routes.

Removed a few unused C++ functions to reduce the size of the DLL.