Various Fixes, No ScratchBuffer Instantiation
The ScratchBuffer is no longer instantiated because it does not need to
be (I thought an old bug was caused by this, but it turned out not be
the case, and I've since fixed that bug).

When calculating the unit's base moves for pathfinding purposes,
retrieving a land unit's regular or embarked movement now always depends
on the originating plot instead of inconsistently depending on the
originating plot or the target plot.

If a unit could walk on a water tile, pathfinders will no longer think
that it is a water tile.

Optimized the way calculated movement rates are handled to reduce double
counting and unnecessary function calls.