Various Fixes, Helpers
A pillaged road improvement will no longer give bonus yields to
improvements in the plot (if there are improvements that would gain
yield bonuses from having a road in the plot).

Added the ability to calculate a plot's yield assuming a working city
(instead of only being able to rely on the current working city).

Potential yield changes from building a road or building an improvement
that gets altered yields from roads are now properly calculated.

Yields changes based on an building in the working city that the
presence of a certain plot feature (eg. Universities' +2 science from
Jungles) are now properly ignored when features would otherwise be
ignored for calculating the yield of a tile after a certain build order.