Ruins Yield Alert, Resolve Working Plot Conflicts
If the player receives some sort of yield from Ancient Ruins, a floating
text pop-up will appear above the plot similar to the ones that appear
when killing a unit under the effects of Honor finisher.

Added a new function that allocates which city should be working plots
that can be worked by multiple cities.