Typo Corrections, Defines Corrections and Update
Fixed a few typos.

Updated certain defines and sections of new code so that their strict
dependence on one another is reduced and/or explicitly defined (thanks
ThorHammerz!).

CvPlayer::IsPlotUnderImmediateThreat() now accepts a single argument as
well, which is the same as calling the PlayerTypes version using the
CvPlayer object's ID.