Binomial RidgeBuilder
CvFractal's RidgeBuilder now uses the binomial RNG to generate voronoi
seeds' weakness and directional bias strength. In the case of weakness,
the non-binomial version had a 50% chance of generating 0, a 16.7%
chance of generating 1, a 16.7% chance of generating 2, and a 16.7%
chance of generating 3; the binomial version has a 50% chance of
generating 0, a 31.25% chance of generating 1, a 15.625% chance of
generating 2, and a 3.125% chance of generating 3, thus smoothing things
out. The situation is roughly the same with directional bias strength:
same 50% chance of 0, but a downward slope of probabilities with
increasing directional bias strength, thus reducing the noisiness of
ridges.