Constification, Pruning, Misc. Optimizations and Fixes
Made certain pointers within functions that pointed to XML entries be
const w.r.t. said entries.

When calculating total yield value, one less call used to check players'
growth thresholds.

Removed some lines of code that were added in previous versions/commits,
but are now redundant and/or unnecessary.

Pruned away the algorithm that prioritizes multiple specialists of the
same type to be in the same building when possible because the effects
are entirely cosmetic despite this bit adding a fair bit to the
function's overhead.

Fixed a possible null pointer dereference if a city has more than one
specialist and does not have a majority religion.