Tweaks toward Safety
Made two tweaks that make the AI be more conservative with city
placement:
1) If the AI has only one city, it will use its reinforce time as both
the closest and second closest reinforce times, instead of using it only
as the closest time and setting the second closest time to 0.
2) Defensive flavor now properly reduces the width of the Gaussian
governing the multiplier for allied reinforce speeds (instead of only
reducing it only if it has a value of 1, as before).