Turn Timer Fixes, Tweaks
Fixed a poor interaction between the WIP new Hybrid mode and turn
timers.

Changed turn timers in the new, WIP Hybrid mode so that they are
calculated at the beginning of a game turn on a per-round basis and
cached for the rest of the game turn in a vector instead of being the
same for all players but having the capability of changing mid-turn.